Align executable emulator and simulator semantics
Implement signed MIN/MAX and arithmetic 2/ in both the assembled BIOS and hosted runtime, correct Python X25519 A24, and make signed division and remainder exact without host floating-point or C++ overflow. Normalize absent optional-memory geometry and retain the generic cyclic NTT and 64-byte ML-KEM seed contracts.

Repair KDOS LOAD, APP-LOAD, and REQUIRE around one checked transaction guard. Every physical line and final compiler state is checked; extent failures and catchable guest throws unwind evaluator state, provisional nested module identities, dictionary HERE/LATEST, allocations, loader frames, diagnostics, and ambient paths before exact rethrow. Buffer constructors now publish the allocator's returned address and startup no longer frees a failed DMA result.

Refresh the ROM, byte-exact source fixtures, source and transport pins, and focused emulator/simulator coverage for arithmetic extremes, X25519 vectors, optional-memory absence, loader rollback, CRLF prescan, nested failures, module transactions, buffer allocation, ring layout, and startup failure propagation.

+++ simulator/core_words.py
@@ -204,18 +204,13 @@ def _negate(context: ExecutionContext) -> None:
 def _minimum(context: ExecutionContext) -> None:
     right = context.data.pop()
     left = context.data.pop()
-    # The BIOS documentation calls this signed, but the current MP64 `cmp`
-    # branch uses the architectural unsigned G/LE conditions.  Preserve the
-    # executable BIOS behavior used by unchanged source.
-    context.data.push(left if left <= right else right)
+    context.data.push(left if s64(left) <= s64(right) else right)
 
 
 def _maximum(context: ExecutionContext) -> None:
     right = context.data.pop()
     left = context.data.pop()
-    # See _minimum: the checked-in BIOS currently compares these cells as
-    # unsigned even though the reference describes signed MIN/MAX.
-    context.data.push(left if left > right else right)
+    context.data.push(left if s64(left) > s64(right) else right)
 
 
 def _one_minus(context: ExecutionContext) -> None:
@@ -231,9 +226,7 @@ def _two_multiply(context: ExecutionContext) -> None:
 
 
 def _two_divide(context: ExecutionContext) -> None:
-    # The executable BIOS uses a logical right shift even though an old source
-    # comment calls 2/ arithmetic.  Preserve that cell-level behavior.
-    context.data.push(context.data.pop() >> 1)
+    context.data.push(s64(context.data.pop()) >> 1)
 
 
 def _and(context: ExecutionContext) -> None:
